--- a/core/src/main/java/org/mule/runtime/core/internal/event/DefaultEventContext.java
+++ b/core/src/main/java/org/mule/runtime/core/internal/event/DefaultEventContext.java
@@ -9,7 +9,6 @@ package org.mule.runtime.core.internal.event;
 import static org.mule.runtime.core.api.util.StringUtils.EMPTY;
 import static org.mule.runtime.core.privileged.event.DefaultFlowCallStack.newDefaultFlowCallStack;
 
-import static java.lang.System.identityHashCode;
 import static java.lang.System.lineSeparator;
 import static java.time.Instant.now;
 import static java.util.Objects.requireNonNull;
@@ -346,9 +345,7 @@ public final class DefaultEventContext extends AbstractEventContext implements S
       this.root = parent.getRootContext();
       this.parent = parent;
       this.componentLocation = componentLocation;
-      this.id = parent.getId() != null
-          ? parent.getId().concat("_").concat(Integer.toString(identityHashCode(this)))
-          : Integer.toString(identityHashCode(this));
+      this.id = parent.nextChildId();
       this.correlationId = correlationId != null ? correlationId : parent.getCorrelationId();
       this.rootId = root.getRootId();
       if (parent instanceof SpanContextAware) {