--- a/core/src/main/java/org/mule/runtime/core/internal/event/DefaultEventContext.java
+++ b/core/src/main/java/org/mule/runtime/core/internal/event/DefaultEventContext.java
@@ -20,6 +20,7 @@ import static org.slf4j.LoggerFactory.getLogger;
 
 import org.mule.runtime.api.component.location.ComponentLocation;
 import org.mule.runtime.api.event.EventContext;
+import org.mule.runtime.api.streaming.Cursor;
 import org.mule.runtime.api.streaming.CursorProvider;
 import org.mule.runtime.api.streaming.bytes.CursorStreamProvider;
 import org.mule.runtime.core.api.construct.FlowConstruct;
@@ -98,8 +99,8 @@ public final class DefaultEventContext extends AbstractEventContext implements S
                                        FlowExceptionHandler exceptionHandler, final String correlationId) {
     BaseEventContext child = new ChildEventContext(parent, componentLocation.orElse(null), exceptionHandler,
                                                    parent.getDepthLevel() + 1, correlationId);
-    if (parent instanceof AbstractEventContext) {
-      ((AbstractEventContext) parent).addChildContext(child);
+    if (parent instanceof AbstractEventContext aec) {
+      aec.addChildContext(child);
     }
 
     return child;
@@ -289,16 +290,14 @@ public final class DefaultEventContext extends AbstractEventContext implements S
    * @return a tracked {@link CursorProvider}.
    * @since 4.3.0
    */
-  public CursorProvider track(ManagedCursorProvider provider, StreamingGhostBuster ghostBuster) {
+  public <T extends Cursor> CursorProvider<T> track(ManagedCursorProvider<T> provider, StreamingGhostBuster ghostBuster) {
     return streamingState.addProvider(provider, ghostBuster);
   }
 
   private void eventContextMaintain(EventContextService eventContextService) {
-    if (eventContextService != null && eventContextService instanceof DefaultEventContextService) {
-      ((DefaultEventContextService) eventContextService).addContext(this);
-      this.onTerminated((e, t) -> {
-        ((DefaultEventContextService) eventContextService).removeContext(DefaultEventContext.this);
-      });
+    if (eventContextService instanceof DefaultEventContextService decs) {
+      decs.addContext(this);
+      this.onTerminated((e, t) -> decs.removeContext(DefaultEventContext.this));
     }
   }
 
@@ -348,8 +347,8 @@ public final class DefaultEventContext extends AbstractEventContext implements S
       this.id = parent.nextChildId();
       this.correlationId = correlationId != null ? correlationId : parent.getCorrelationId();
       this.rootId = root.getRootId();
-      if (parent instanceof SpanContextAware) {
-        SpanContext parentSpanContext = getParentDistributedTraceContext((SpanContextAware) parent);
+      if (parent instanceof SpanContextAware sca) {
+        SpanContext parentSpanContext = getParentDistributedTraceContext(sca);
         if (parentSpanContext != null) {
           spanContext = parentSpanContext.copy();
         }
@@ -358,11 +357,6 @@ public final class DefaultEventContext extends AbstractEventContext implements S
 
     private SpanContext getParentDistributedTraceContext(SpanContextAware parent) {
       return parent.getSpanContext();
-    }
-
-    private ChildEventContext(BaseEventContext parent, ComponentLocation componentLocation,
-                              FlowExceptionHandler messagingExceptionHandler, int depthLevel) {
-      this(parent, componentLocation, messagingExceptionHandler, depthLevel, null);
     }
 
     @Override