--- a/core/src/main/java/org/mule/runtime/core/internal/streaming/EventStreamingState.java
+++ b/core/src/main/java/org/mule/runtime/core/internal/streaming/EventStreamingState.java
@@ -6,18 +6,23 @@
  */
 package org.mule.runtime.core.internal.streaming;
 
-import static java.lang.System.identityHashCode;
-import static org.slf4j.LoggerFactory.getLogger;
 import static org.mule.runtime.core.internal.streaming.CursorManager.STREAMING_VERBOSE;
 import static org.mule.runtime.core.internal.streaming.CursorUtils.unwrap;
+
+import static java.lang.System.identityHashCode;
+
+import static org.slf4j.LoggerFactory.getLogger;
+
+import org.mule.runtime.api.component.location.ComponentLocation;
+import org.mule.runtime.api.streaming.Cursor;
+import org.mule.runtime.api.streaming.CursorProvider;
 
 import java.lang.ref.WeakReference;
 import java.util.Optional;
 
 import com.github.benmanes.caffeine.cache.Cache;
 import com.github.benmanes.caffeine.cache.Caffeine;
-import org.mule.runtime.api.component.location.ComponentLocation;
-import org.mule.runtime.api.streaming.CursorProvider;
+
 import org.slf4j.Logger;
 
 /**
@@ -40,9 +45,10 @@ public class EventStreamingState {
    * @param ghostBuster the {@link StreamingGhostBuster} used to do early reclamation of the {@code provider}
    * @return the {@link ManagedCursorProvider} that must continue to be used
    */
-  public ManagedCursorProvider addProvider(ManagedCursorProvider provider, StreamingGhostBuster ghostBuster) {
+  public <T extends Cursor> ManagedCursorProvider<T> addProvider(ManagedCursorProvider<T> provider,
+                                                                 StreamingGhostBuster ghostBuster) {
     final int id = provider.getId();
-    ManagedCursorProvider managedProvider = getOrAddManagedProvider(id, provider, ghostBuster);
+    ManagedCursorProvider<T> managedProvider = getOrAddManagedProvider(id, provider, ghostBuster);
 
     // This can happen when a foreach component splits a text document using a stream.
     // Iteration N might try to manage the same root provider that was already managed in iteration N-1, but the
@@ -61,12 +67,12 @@ public class EventStreamingState {
     return managedProvider;
   }
 
-  private ManagedCursorProvider getOrAddManagedProvider(int id,
-                                                        ManagedCursorProvider provider,
-                                                        StreamingGhostBuster ghostBuster) {
+  private <T extends Cursor> ManagedCursorProvider<T> getOrAddManagedProvider(int id,
+                                                                              ManagedCursorProvider<T> provider,
+                                                                              StreamingGhostBuster ghostBuster) {
     return providers.get(id, k -> {
       if (STREAMING_VERBOSE) {
-        CursorProvider innerDelegate = unwrap(provider);
+        CursorProvider<T> innerDelegate = unwrap(provider);
         Optional<ComponentLocation> originatingLocation = provider.getOriginatingLocation();
         LOGGER.info("Added ManagedCursorProvider: {} for delegate: {} opened by: {}", k, identityHashCode(innerDelegate),
                     originatingLocation.map(ComponentLocation::getLocation).orElse("unknown"));
@@ -80,7 +86,7 @@ public class EventStreamingState {
    */
   public void dispose() {
     providers.asMap().forEach((hash, weakReference) -> {
-      ManagedCursorProvider provider = weakReference.get();
+      ManagedCursorProvider<?> provider = weakReference.get();
       if (provider != null) {
         weakReference.clear();
         provider.releaseResources();