--- a/core/src/main/java/org/mule/runtime/core/privileged/event/BaseEventContext.java
+++ b/core/src/main/java/org/mule/runtime/core/privileged/event/BaseEventContext.java
@@ -202,4 +202,11 @@ public interface BaseEventContext extends EventContext {
    */
   int getDepthLevel();
 
+  /**
+   * @return a unique id with the if of this context as prefix
+   * 
+   * @since 4.10, 4.9.4, 4.6.16, 4.4.0-202505
+   */
+  String nextChildId();
+
 }