… modules/extensions-support/src/test/java/org/mule/runtime/module/extension/internal/runtime/execution/SdkInternalContextTestCase.java
… tests/performance/pom.xml
… tests/performance/src/main/java/org/mule/runtime/module/extension/internal/runtime/execution/SdkInternalContextBenchmark.java
… tests/performance/src/test/java/org/mule/test/performance/util/AbstractIsolatedFunctionalPerformanceTestCase.java