Improvements in times:

```Benchmark                                                                    Mode  Cnt        Score       Error   Units
Baseline:
EventContextBenchmark.createEventContextWith1000ChildrenTerminateAllAtOnce   avgt   10    378451.152 ± 16140.772     ns/op
SdkInternalContextBenchmark.contextWith300ParallelChildren                   avgt   10  14594822.655 ± 1281822.849   ns/op
SdkInternalContextBenchmark.contextWith3Parallel                             avgt   10     90799.671 ±    3848.457   ns/op

After:
EventContextBenchmark.createEventContextWith1000ChildrenTerminateAllAtOnce   avgt   10    342399.367 ± 28087.914    
 ns/op (-9.5%)
SdkInternalContextBenchmark.contextWith300ParallelChildren                   avgt   10   1813778.934 ± 173664.899    ns/op (-87,5%)
SdkInternalContextBenchmark.contextWith3Parallel                             avgt   10     37869.280 ±   3338.925    ns/op (-58,3%)
```